+++ src/barch.cpp
@@ -268,7 +268,7 @@ int KEYS(caller& call, const arg_t& argv) {
     art::value_type pattern = cpat;
     if (spec.count) {
         for (auto shard : barch::get_shard_count()) {
-            call.kspace()->get(shard)->glob(spec, pattern, [&](const art::leaf & unused(l)) -> bool {
+            call.kspace()->get(shard)->glob(spec, pattern, false, [&](const art::leaf & unused(l)) -> bool {
                 ++replies;
                 return true;
             });
@@ -279,7 +279,7 @@ int KEYS(caller& call, const arg_t& argv) {
         call.start_array();
 
         for (auto shard : barch::get_shard_count()) {
-            call.kspace()->get(shard)->glob(spec, pattern, [&](const art::leaf &l) -> bool {
+            call.kspace()->get(shard)->glob(spec, pattern, false, [&](const art::leaf &l) -> bool {
                 std::lock_guard lk(vklock); // because there's worker threads concurrently calling here
                 if (0 != call.push_encoded_key(l.get_key())) {
                     return false;
@@ -316,7 +316,7 @@ int VALUES(caller& call, const arg_t& argv) {
     art::value_type pattern = cpat;
     if (spec.count) {
         for (auto shard : barch::get_shard_count()) {
-            call.kspace()->get(shard)->glob(spec, pattern, [&](const art::leaf & unused(l)) -> bool {
+            call.kspace()->get(shard)->glob(spec, pattern, true, [&](const art::leaf & unused(l)) -> bool {
                 ++replies;
                 return true;
             });
@@ -327,7 +327,7 @@ int VALUES(caller& call, const arg_t& argv) {
         call.start_array();
 
         for (auto shard : barch::get_shard_count()) {
-            call.kspace()->get(shard)->glob(spec, pattern, [&](const art::leaf &l) -> bool {
+            call.kspace()->get(shard)->glob(spec, pattern, true, [&](const art::leaf &l) -> bool {
                 std::lock_guard lk(vklock); // because there's worker threads concurrently calling here
                 if (0 != call.push_encoded_key(l.get_key())) {
                     return false;